need to move these files over to another computer. dont think they should get in anyone else's way

+++ trunk/desmume/src/memorystream.h
@@ -215,7 +215,7 @@ protected:
 
 	int overflow(int c)
 	{
-		expand(-1);
+		expand((size_t)-1);
 		dosync(c);
 		return 1;
 	}
